Use the function directly
